\name{muc}
\alias{muc}
\docType{data}
\title{Object of class osmar from central Munich}
\description{
Data retrieved with \code{get_osm(center_bbox(11.575278, 48.137222, 200, 200))}.
}
\usage{data(muc)}
\format{
  The format is:
List of 3
 $ nodes    :List of 2
  ..$ attrs:'data.frame':	975 obs. of  9 variables:
  .. ..$ id       : num [1:975] 1955016 17780035 18929510 18929515 18929522 ...
  .. ..$ lat      : num [1:975] 48.1 48.1 48.1 48.1 48.1 ...
  .. ..$ lon      : num [1:975] 11.6 11.6 11.6 11.6 11.6 ...
  .. ..$ user     : Factor w/ 36 levels "chan","ckol",..: 26 24 13 12 21 6 6 13 26 21 ...
  .. ..$ uid      : Factor w/ 36 levels "107037","109029",..: 6 29 14 34 15 10 10 14 6 15 ...
  .. ..$ visible  : Factor w/ 1 level "true": 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ version  : num [1:975] 3 35 3 3 6 6 6 3 3 3 ...
  .. ..$ changeset: num [1:975] 10239803 10484152 6909578 1460631 10162612 ...
  .. ..$ timestamp: POSIXlt[1:975], format: "2011-12-29 21:07:53" "2012-01-24 12:51:04" ...
  ..$ tags :'data.frame':	662 obs. of  3 variables:
  .. ..$ id: num [1:662] 17780035 17780035 17780035 17780035 17780035 ...
  .. ..$ k : Factor w/ 109 levels "addr:city","addr:country",..: 18 35 36 37 42 43 44 45 46 47 ...
  .. ..$ v : Factor w/ 291 levels "-0.5","-1","-2",..: 43 196 211 102 194 167 288 170 194 196 ...
  ..- attr(*, "class")= chr [1:3] "nodes" "osmar_element" "list"
 $ ways     :List of 3
  ..$ attrs:'data.frame':	214 obs. of  7 variables:
  .. ..$ id       : num [1:214] 9.66e+07 8.58e+07 8.58e+07 1.05e+08 1.05e+08 ...
  .. ..$ user     : Factor w/ 26 levels "FK270673","FloSch",..: 22 7 7 23 22 8 23 19 21 11 ...
  .. ..$ uid      : Factor w/ 26 levels "109029","130472",..: 24 10 10 15 24 6 15 22 3 9 ...
  .. ..$ visible  : Factor w/ 1 level "true": 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ version  : num [1:214] 2 1 1 2 1 3 2 2 2 3 ...
  .. ..$ changeset: num [1:214] 7622488 6411339 6411339 7857000 7622488 ...
  .. ..$ timestamp: POSIXlt[1:214], format: "2011-03-20 22:13:47" "2010-11-20 00:18:02" ...
  ..$ tags :'data.frame':	607 obs. of  3 variables:
  .. ..$ id: num [1:607] 9.66e+07 9.66e+07 8.58e+07 8.58e+07 1.05e+08 ...
  .. ..$ k : Factor w/ 52 levels "addr:city","addr:country",..: 22 26 11 11 25 26 22 26 11 25 ...
  .. ..$ v : Factor w/ 88 levels "-0.5","-1","-2",..: 64 6 88 88 4 4 44 4 88 4 ...
  ..$ refs :'data.frame':	1262 obs. of  2 variables:
  .. ..$ id : num [1:1262] 96619179 96619179 85765758 85765758 85765758 ...
  .. ..$ ref: num [1:1262] 1.12e+09 3.40e+08 9.96e+08 9.96e+08 9.96e+08 ...
  ..- attr(*, "class")= chr [1:3] "ways" "osmar_element" "list"
 $ relations:List of 3
  ..$ attrs:'data.frame':	56 obs. of  7 variables:
  .. ..$ id       : num [1:56] 1773072 1796136 1843975 1792663 30479 ...
  .. ..$ user     : Factor w/ 20 levels "Andreas Binder",..: 15 8 8 8 12 10 12 9 15 2 ...
  .. ..$ uid      : Factor w/ 20 levels "109029","13832",..: 1 16 16 16 2 14 2 8 1 6 ...
  .. ..$ visible  : Factor w/ 1 level "true": 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ version  : num [1:56] 14 5 6 13 91 4 48 7 15 62 ...
  .. ..$ changeset: num [1:56] 10510995 10210507 10210507 10210507 10393071 ...
  .. ..$ timestamp: POSIXlt[1:56], format: "2012-01-27 10:36:50" "2011-12-26 19:56:40" ...
  ..$ tags :'data.frame':	425 obs. of  3 variables:
  .. ..$ id: num [1:425] 1773072 1773072 1773072 1773072 1773072 ...
  .. ..$ k : Factor w/ 47 levels "admin_level",..: 10 14 24 25 26 33 35 43 44 46 ...
  .. ..$ v : Factor w/ 150 levels "-3","0","09",..: 93 106 91 70 40 125 139 61 100 72 ...
  ..$ refs :'data.frame':	6119 obs. of  4 variables:
  .. ..$ id  : num [1:6119] 1773072 1773072 1773072 1773072 1773072 ...
  .. ..$ type: Factor w/ 3 levels "node","relation",..: 1 3 1 3 1 3 1 3 3 1 ...
  .. ..$ ref : num [1:6119] 1.45e+09 1.32e+08 6.00e+07 5.59e+07 6.00e+07 ...
  .. ..$ role: Factor w/ 11 levels "","admin_centre",..: 11 8 11 8 11 8 11 9 10 11 ...
  ..- attr(*, "class")= chr [1:3] "relations" "osmar_element" "list"
 - attr(*, "class")= chr [1:2] "osmar" "list"
}
\source{\url{http://www.openstreetmap.org/}, downloaded 10 February 2012.}
\examples{
data("muc", package = "osmar")
}
\seealso{
  \code{\link{find}}, \code{\link{as_sp}}
}
\keyword{datasets}
